Attempting to move the C library into an optional dependency so it can be disabled.
