+++ Cargo.toml
@@ -32,13 +32,19 @@ semver = "0.9.0"
 neon-runtime = { version = "=0.3.1", path = "crates/neon-runtime" }
 
 [features]
-default = []
-n-api = ["neon-runtime/n-api", "neon-build/n-api"]
+default = ["legacy-runtime"]
+
+# Feature flag to enable the legacy V8/NAN runtime. For now, this feature is
+# enabled by default. Disabling the feature uses the experimental N-API runtime
+# instead. If the N-API works out, once it's complete we'll be able to turn this
+# flag off by default and eventually remove the legacy runtime entirely.
+legacy-runtime = ["neon-runtime/neon-sys"]
 
 [workspace]
 members = [
 	"crates/neon-build",
 	"crates/neon-runtime",
+    "crates/neon-sys",
 	"test/static",
 	"test/electron/native",
 	"test/dynamic/native",
