--- crates/neon-runtime/native-napi/binding.gyp
@@ -1,8 +1,0 @@
-{
-  "targets": [
-    {
-      "target_name": "neon",
-      "sources": [ "src/neon.cc" ]
-    }
-  ]
-}
