--- crates/neon-runtime/native-napi/src/neon.cc
@@ -1,2 +1,0 @@
-#include <node_api.h>
-#include "neon.h"
