+++ crates/neon-runtime/src/lib.rs
@@ -3,11 +3,11 @@ extern crate cfg_if;
 use cfg_if::cfg_if;
 
 cfg_if! {
-    if #[cfg(feature =  "n-api")] {
-        mod napi;
-        pub use napi::*;
-    } else {
+    if #[cfg(feature = "neon-sys")] {
         mod nan;
         pub use nan::*;
+    } else {
+        mod napi;
+        pub use napi::*;
     }
 }
