+++ crates/neon-sys/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "neon-sys"
+version = "0.3.1"
+authors = ["David Herman <david.herman@gmail.com>"]
+description  = "Exposes the low-level V8/NAN C/C++ APIs. Will be superseded by N-API."
+edition = "2018"
+license = "MIT/Apache-2.0"
+links = "neon"     # libneon.{a,obj} gets linked in with this package
+build = "build.rs" # this script builds libneon.{a,obj}
+
+[dependencies]
+
+[build-dependencies]
+cc = "1.0.0"
+regex = "1.0.0"
