+++ crates/neon-sys/native/src/neon_string.h
